Roaming behaviour bug fix, stripped more unused prototype assets
Fixed bug: The monster when leaving vents to search the player location will proceed to check around the area a few times. Unfortunately it updated the area to search each time making it impossible for the player to disengage. The area to search now updates once and is reused for search attempts.