Introduced Director AI
Director will gradually replace all Timed functions thus make it's client Monster simpler to debug. As well as allowing the Monster itself to handle more tasks.

So far Menace Gauge implementation has begun to work and only needs to be expanded upon through further testing.

Venting behaviour will need to be reworked entirely for the Director to push it in and poke it out.